MockSamples moved to mock.py, will be some broken tests I still need to fix...
